refactor(core): extract search/embedding bootstrap slice (#214)
Made-with: Cursor
